Elusive Error Struct Bug
Tracking an elusive error struct bug: &errors.errorString{s:””}. For
some reason, the struct is seen as non-nil when returned from a package
or certain functions and it triggers an error condition, even though
the original function didn’t return an actual error. Workaround hack is
to use err.Error().